pyproject.toml
[project]
name = "crypto-trailing-stop"
version = "0.9.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "José María Sola Durán", email = "josemaria.sola.duran@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "apscheduler>=3.11.0",
    "authlib>=1.6.0",
    "cachebox>=5.0.1",
    "ccxt>=4.4.91",
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "piccolo[sqlite]>=1.27.1",
    "pydantic-settings>=2.8.1",
    "pydash>=8.0.5",
    "ta>=0.11.0",
    "uvicorn>=0.34.0",
]

[tool.taskipy.tasks]
start = "fastapi dev src/crypto_trailing_stop/main.py --app app --reload --host 0.0.0.0"
test = "pytest -x --log-cli-level=INFO"
"test:coverage" = "pytest --cov=./src --log-cli-level=INFO"
"test:coverage:html" = "pytest --cov=./src --cov-report=html --log-cli-level=INFO"
lint = "pre-commit install && pre-commit run --all"


[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "faker>=37.1.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.2.1",
    "pytest-httpserver>=1.1.3",
    "python-dotenv>=1.1.0",
    "taskipy>=1.14.1",
    "testcontainers>=4.10.0",
]

[project.scripts]
crypto-trailing-stop = "crypto_trailing_stop:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = [
    "I",   # isort
    "UP",  # pyupgrade
    "E501", # Add the `line-too-long` rule to the enforced rule set.
    "T201" # Remove print statements
]
ignore = ["UP038"]

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 10.
max-complexity = 10

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.pytest.ini_options]
log_cli_level = "INFO"
asyncio_mode = "auto"
python_functions = "should_* test_*"

[tool.coverage.run]
source = ["src"]
branch = true  # optional, but better for full logic path coverage
concurrency = ["thread", "multiprocessing"]

[tool.coverage.report]
show_missing = true
# fail_under = 80
